refactor: 사용자, 레이블, 마일스톤 목록 조회 로직을 각 domain 패키지 하위에 이동하여 구현
