+++ backend/src/main/java/codesquad/team01/issuetracker/user/dto/UserDto.java
@@ -19,6 +19,7 @@ public class UserDto {
 	public record WriterResponse(
 		// 처음엔 작성자 응답 DTO로 생성했지만
 		// 현재 담당자, 작성자 필터 db 조회용 dto, 응답용 dto의 역할도 하고 있음
+		// 이럴 경우 더 만들지? 이름을 범용적으로 바꿀지?
 		Integer id,
 
 		String username,
@@ -45,7 +46,7 @@ public class UserDto {
 	public record UserFilterListResponse(
 		int totalCount,
 		List<WriterResponse> users // WriterResponse와 같이 데이터를 필요로 해서 임시로 넣음
-		// 이럴 경우 하나 더 만들지? 이름을 범용적으로 바꿀지?
+
 	) {
 	}
